% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_bruteforce.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute forcing knapsack problems is used for finding all the possible combinations of given data.}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{This argument is contain a data frame of value and weight of items}

\item{W}{This is the capacity of knapsack}

\item{parallel}{This parameter is to run some code in parallel core or not}
}
\value{
This return the list contain value and elements
}
\description{
Brute forcing knapsack problems is used for finding all the possible combinations of given data.
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
data.frame(w=sample(1:4000, size = n, replace = TRUE),
v=runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
}
