% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic problems is used for finding all the possible combinations of given data dynamically.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{This argument is contain a data frame of value and weight of items}

\item{W}{This is the capacity of knapsack}
}
\value{
This return the list contain value and elements
}
\description{
knapsack_dynamic problems is used for finding all the possible combinations of given data dynamically.
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
data.frame(w=sample(1:4000, size = n, replace = TRUE),
v=runif(n = n, 0, 10000))
knapsack_dynamic(x = knapsack_objects[1:12,], W = 3500)
}
